# contracts/admin-api.openapi.yaml
openapi: 3.1.0
info:
  title: AI Control Layer – Admin API
  description: Management and reporting API consumed by the admin panel.
  version: 1.0.0
  x-gateway-version: 0.1.0
paths:
...
  /admin/v1/events:
    get:
      tags:
      - events
      summary: List Events
      operationId: listEvents
      security:
      - HTTPBearer: []
      parameters:
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
      - name: until
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Until
      - name: subject
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
      - name: group
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Group
      - name: agent
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent
      - name: control
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Control
      - name: rule_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Rule Id
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Action'
          - type: 'null'
          title: Action
      - name: point
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/InspectionPoint'
          - type: 'null'
          title: Point
      - name: event_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Event Type
      - name: taxonomy
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Taxonomy
      - name: before_seq
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Before Seq
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventSummary'
                title: Response Listevents
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/events/stream:
    get:
      tags:
      - events
      summary: Stream Events
      operationId: streamEvents
      security:
      - HTTPBearer: []
      parameters:
      - name: subject
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Action'
          - type: 'null'
          title: Action
      responses:
        '200':
          description: Server-sent events; each `data:` line is an EventSummary JSON object.
          content:
            application/json:
              schema: {}
            text/event-stream: {}
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/events/{event_id}:
    get:
      tags:
      - events
      summary: Get Event
      description: Full decision trace (the audit record) for one event.
      operationId: getEvent
      security:
      - HTTPBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvent'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/incidents:
    get:
      tags:
      - events
      summary: List Incidents
      operationId: listIncidents
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incident'
                title: Response Listincidents
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/incidents/{incident_id}:
    get:
      tags:
      - events
      summary: Get Incident
      operationId: getIncident
      security:
      - HTTPBearer: []
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - events
      summary: Update Incident
      operationId: updateIncident
      security:
      - HTTPBearer: []
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/audit/verify:
    get:
      tags:
      - events
      summary: Verify Chain
      operationId: verifyAuditChain
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainVerifyResult'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/audit/export:
    get:
      tags:
      - events
      summary: Export Audit
      operationId: exportAudit
      security:
      - HTTPBearer: []
      parameters:
      - name: format
        in: query
        required: false
        schema:
          enum:
          - jsonl
          - ocsf
          - csv
          type: string
          default: jsonl
          title: Format
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
      - name: until
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Until
      responses:
        '200':
          description: JSONL (raw records), OCSF JSONL, or CSV.
          content:
            application/json:
              schema: {}
            application/x-ndjson: {}
            text/csv: {}
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Action:
      type: string
      enum:
      - allow
      - monitor
      - redact
      - pseudonymise
      - sanitize
      - route_local
      - downgrade
      - require_approval
      - block
      title: Action
      description: Decision actions (concept §6.3).
...
    ApprovalRef:
      properties:
        approval_id:
          type: string
          title: Approval Id
        status:
          $ref: '#/components/schemas/ApprovalStatus'
          default: pending
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        approver_scope:
          type: string
          title: Approver Scope
          default: admin
          examples:
          - user
          - admin
      additionalProperties: false
      type: object
      required:
      - approval_id
      title: ApprovalRef
...
    AuditDecision:
      properties:
        action:
          $ref: '#/components/schemas/Action'
        applied:
          items:
            $ref: '#/components/schemas/Action'
          type: array
          title: Applied
        would_action:
          anyOf:
          - $ref: '#/components/schemas/Action'
          - type: 'null'
        final:
          type: boolean
          title: Final
          default: false
        decided_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Decided By
        decided_phase:
          anyOf:
          - $ref: '#/components/schemas/Phase'
          - type: 'null'
        rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: Rule Ids
        risk_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Risk Score
          default: 0.0
        reason:
          type: string
          title: Reason
          default: ''
      additionalProperties: false
      type: object
      required:
      - action
      title: AuditDecision
    AuditEvent:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        event_id:
          type: string
          title: Event Id
        seq:
          type: integer
          minimum: 0.0
          title: Seq
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        event_type:
          $ref: '#/components/schemas/EventType'
        severity:
          $ref: '#/components/schemas/Severity'
          default: info
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
        principal:
          anyOf:
          - $ref: '#/components/schemas/AuditPrincipal'
          - type: 'null'
        client:
          anyOf:
          - $ref: '#/components/schemas/ClientInfo'
          - type: 'null'
        point:
          anyOf:
          - $ref: '#/components/schemas/InspectionPoint'
          - type: 'null'
        tool:
          anyOf:
          - type: string
          - type: 'null'
          title: Tool
        server:
          anyOf:
          - type: string
          - type: 'null'
          title: Server
        model_requested:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Requested
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        connector:
          anyOf:
          - type: string
          - type: 'null'
          title: Connector
        args_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Args Hash
          description: Salted hash of canonical tool arguments.
        payload_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Payload Hash
          description: Salted hash of the inspected payload.
        response_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Response Hash
        decision:
          anyOf:
          - $ref: '#/components/schemas/AuditDecision'
          - type: 'null'
        verdicts:
          items:
            $ref: '#/components/schemas/AuditVerdict'
          type: array
          title: Verdicts
        risk_factors:
          items:
            $ref: '#/components/schemas/RiskFactor'
          type: array
          title: Risk Factors
        route:
          anyOf:
          - $ref: '#/components/schemas/RouteInfo'
          - type: 'null'
        approval:
          anyOf:
          - $ref: '#/components/schemas/ApprovalRef'
          - type: 'null'
        labels_after:
          anyOf:
          - $ref: '#/components/schemas/SessionLabels'
          - type: 'null'
        versions:
          $ref: '#/components/schemas/Versions'
        taxonomy:
          $ref: '#/components/schemas/Taxonomy'
        usage:
          anyOf:
          - $ref: '#/components/schemas/Usage'
          - type: 'null'
        latency:
          anyOf:
          - $ref: '#/components/schemas/LatencyBreakdown'
          - type: 'null'
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
        redacted_payload:
          anyOf:
          - type: string
          - type: 'null'
          title: Redacted Payload
          description: Redacted/pseudonymised text only; present only if policy enables it.
        detail:
          additionalProperties: true
          type: object
          title: Detail
          description: Event-type-specific details (no raw sensitive values).
        prev_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Prev Hash
        hash:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Hash
      additionalProperties: false
      type: object
      required:
      - event_id
      - seq
      - timestamp
      - event_type
      - versions
      - prev_hash
      - hash
      title: AuditEvent
...
    AuditPrincipal:
      properties:
        subject:
          type: string
          title: Subject
        kind:
          $ref: '#/components/schemas/PrincipalKind'
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        groups:
          items:
            type: string
          type: array
          title: Groups
        agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        auth_method:
          $ref: '#/components/schemas/AuthMethod'
          default: none
        api_key_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Key Id
        delegation_chain:
          items:
            type: string
          type: array
          title: Delegation Chain
      additionalProperties: false
      type: object
      required:
      - subject
      - kind
      title: AuditPrincipal
    AuditVerdict:
      properties:
        control_id:
          type: string
          title: Control Id
        control_type:
          type: string
          title: Control Type
        phase:
          $ref: '#/components/schemas/Phase'
        action:
          $ref: '#/components/schemas/Action'
        final:
          type: boolean
          title: Final
          default: false
        score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Score
        status:
          $ref: '#/components/schemas/VerdictStatus'
          default: ok
        latency_ms:
          type: number
          title: Latency Ms
          default: 0.0
        rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: Rule Ids
        findings:
          items:
            $ref: '#/components/schemas/AuditFinding'
          type: array
          title: Findings
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      additionalProperties: false
      type: object
      required:
      - control_id
      - control_type
      - phase
      - action
      title: AuditVerdict
    AuthMethod:
      type: string
      enum:
      - jwt
      - api_key
      - client_credentials
      - none
      title: AuthMethod
...
    ChainVerifyResult:
      properties:
        ok:
          type: boolean
          title: Ok
        records:
          type: integer
          title: Records
        head_seq:
          anyOf:
          - type: integer
          - type: 'null'
          title: Head Seq
        head_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Head Hash
        first_bad_seq:
          anyOf:
          - type: integer
          - type: 'null'
          title: First Bad Seq
        message:
          type: string
          title: Message
          default: ''
      additionalProperties: false
      type: object
      required:
      - ok
      - records
      title: ChainVerifyResult
    ClientInfo:
      properties:
        app:
          type: string
          title: App
          default: unknown
          examples:
          - opencode
          - librechat
          - sdk
          - agent
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      additionalProperties: false
      type: object
      title: ClientInfo
...
    EventSummary:
      properties:
        event_id:
          type: string
          title: Event Id
        seq:
          type: integer
          title: Seq
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        event_type:
          type: string
          title: Event Type
        severity:
          $ref: '#/components/schemas/Severity'
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        groups:
          items:
            type: string
          type: array
          title: Groups
        agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
        point:
          anyOf:
          - $ref: '#/components/schemas/InspectionPoint'
          - type: 'null'
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        tool:
          anyOf:
          - type: string
          - type: 'null'
          title: Tool
        action:
          anyOf:
          - $ref: '#/components/schemas/Action'
          - type: 'null'
        rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: Rule Ids
        risk_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Risk Score
        latency_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Latency Ms
      additionalProperties: false
      type: object
      required:
      - event_id
      - seq
      - timestamp
      - event_type
      - severity
      title: EventSummary
    EventType:
      type: string
      enum:
      - decision
      - policy_change
      - policy_reload_failed
      - grant_change
      - incident
      - approval
      - breakglass
      - feed_update
      - feed_verify_failed
      - artifact_scan
      - mcp_drift
      - budget_breach
      - auth_failure
      - system_alert
      title: EventType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Incident:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        category:
          type: string
          title: Category
          examples:
          - forbidden_model
          - mcp_rug_pull
          - exfiltration_attempt
          - budget_breach
        severity:
          $ref: '#/components/schemas/Severity'
        status:
          type: string
          enum:
          - open
          - triaged
          - resolved
          - false_positive
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        event_ids:
          items:
            type: string
          type: array
          title: Event Ids
        rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: Rule Ids
        notes:
          items:
            $ref: '#/components/schemas/IncidentNote'
          type: array
          title: Notes
        detail:
          additionalProperties: true
          type: object
          title: Detail
          description: e.g. rug-pull description diff.
      additionalProperties: false
      type: object
      required:
      - id
      - title
      - category
      - severity
      - status
      - created_at
      - updated_at
      title: Incident
...
    IncidentPatch:
      properties:
        status:
          anyOf:
          - type: string
            enum:
            - open
            - triaged
            - resolved
            - false_positive
          - type: 'null'
          title: Status
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
        note:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      title: IncidentPatch
...
    InspectionPoint:
      type: string
      enum:
      - ingress
      - egress
      - tool_call
      - tool_result
      - embeddings
      - agent_message
      - artifact_load
      - mcp_initialize
      - mcp_tools_list
      title: InspectionPoint
      description: Where in the traffic flow a payload is inspected (concept §6.1).
...
    LatencyBreakdown:
      properties:
        total_ms:
          type: number
          title: Total Ms
          default: 0.0
        pipeline_ms:
          type: number
          title: Pipeline Ms
          default: 0.0
        upstream_ms:
          type: number
          title: Upstream Ms
          default: 0.0
      additionalProperties: false
      type: object
      title: LatencyBreakdown
...
    PrincipalKind:
      type: string
      enum:
      - user
      - agent
      - service
      - anonymous
      title: PrincipalKind
...
    RiskFactor:
      properties:
        factor:
          $ref: '#/components/schemas/RiskFactorName'
        value:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Value
        weight:
          type: number
          minimum: 0.0
          title: Weight
          default: 1.0
        contribution:
          type: number
          minimum: 0.0
          title: Contribution
          description: weight × value, after normalisation.
          default: 0.0
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          description: Control id that produced the factor.
      additionalProperties: false
      type: object
      required:
      - factor
      - value
      title: RiskFactor
...
    RouteInfo:
      properties:
        model_requested:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Requested
        model:
          type: string
          title: Model
          description: Resolved public model id, e.g. `local/general`.
        connector:
          type: string
          title: Connector
        tier:
          $ref: '#/components/schemas/ConnectorTier'
        reason:
          type: string
          title: Reason
          examples:
          - 'auto → local/loan-memo-pl: task=loan_memo (0.91), data=confidential → local'
        degraded:
          type: boolean
          title: Degraded
          default: false
        factors:
          additionalProperties: true
          type: object
          title: Factors
      additionalProperties: false
      type: object
      required:
      - model
      - connector
      - tier
      - reason
      title: RouteInfo
    SessionLabels:
      properties:
        integrity:
          $ref: '#/components/schemas/Integrity'
          default: trusted
        confidentiality:
          $ref: '#/components/schemas/DataClass'
          default: public
        taint:
          items:
            $ref: '#/components/schemas/TaintFlag'
          type: array
          title: Taint
        sources:
          items:
            type: string
          type: array
          title: Sources
          description: Rule ids that raised the labels.
      additionalProperties: false
      type: object
      title: SessionLabels
      description: IFC labels carried by a session (concept §9). Only ever raised without approval.
    Severity:
      type: string
      enum:
      - info
      - low
      - medium
      - high
      - critical
      title: Severity
...
    Taxonomy:
      properties:
        owasp_llm:
          items:
            type: string
          type: array
          title: Owasp Llm
          examples:
          - - LLM01:2025
        owasp_agentic:
          items:
            type: string
          type: array
          title: Owasp Agentic
          examples:
          - - ASI01
        owasp_mcp:
          items:
            type: string
          type: array
          title: Owasp Mcp
          examples:
          - - MCP03
        atlas:
          items:
            type: string
          type: array
          title: Atlas
          examples:
          - - AML.T0051
        cve:
          items:
            type: string
          type: array
          title: Cve
      additionalProperties: false
      type: object
      title: Taxonomy
      description: Framework tags attached to rules and events (concept §13).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    Versions:
      properties:
        policy:
          type: string
          title: Policy
          description: Compiled policy version id.
        grants:
          type: string
          title: Grants
          description: Grants-table version (monotonic).
          default: '0'
        signatures:
          type: string
          title: Signatures
          description: Signature bundle version.
          default: '0'
        classifiers:
          additionalProperties:
            type: string
          type: object
          title: Classifiers
        judges:
          additionalProperties:
            type: string
          type: object
          title: Judges
        gateway:
          type: string
          title: Gateway
          default: 0.1.0
      additionalProperties: false
      type: object
      required:
      - policy
      title: Versions
      description: Everything needed to replay a decision (concept §13).
  securitySchemes:
    HTTPBearer:
      type: http
      description: Keycloak access token (JWT) or personal API key (`acl_...`).
      scheme: bearer
servers:
- url: http://localhost:8000